Supporting a 64 bit build on Windows now
Refactored a bit of the CMake project to support an explicit directory for
the Bin directory of GLEW and GLFW. This makes it a bit more clear what
directories need to be pointed to where and when.

Apparantly in x64 builds on Windows std::vector returns size() as size_t
rather than unsigned int. Hopefully this doesn't mess with other
platforms. Maybe g++ was just auto-casting while MSVC likes to throw
warnings. size_t is now used for a bunch of internal size-related values

+++ BrickwareGraphics/include/BrickwareGraphics/Mesh.hpp
@@ -62,11 +62,11 @@ namespace Brickware
 
 			Math::Bounds getBounds();
 
-			int getPointSize();
-			int getNormalSize();
-			int getIndexSize();
-			int getNumberOfVerts();
-			int getTexCoordSize();
+			size_t getPointSize();
+			size_t getNormalSize();
+			size_t getIndexSize();
+			size_t getNumberOfVerts();
+			size_t getTexCoordSize();
 
 			void setBufferHint(BufferHint hint);
 
@@ -115,13 +115,13 @@ namespace Brickware
 
 			Math::Bounds bounds;
 
-			unsigned int numberOfVerts;
+			size_t numberOfVerts;
 
-			unsigned int pointSize;
-			unsigned int normalSize;
-			unsigned int texCoordSize;
+			size_t pointSize;
+			size_t normalSize;
+			size_t texCoordSize;
 
-			unsigned int indexSize;
+			size_t indexSize;
 
 #ifdef GL_SUPPORT
 			//Buffer IDs
